Moved XXXAnalyzer.spec.ts to ./analyzer subfolder.
