+++ tests/unit/utils/analyzer/BalanceAnalyzer.spec.ts
@@ -25,7 +25,7 @@ import {flushPromises} from "@vue/test-utils";
 import MockAdapter from "axios-mock-adapter";
 import axios from "axios";
 import {BalanceAnalyzer} from "@/utils/analyzer/BalanceAnalyzer";
-import {SAMPLE_ACCOUNT, SAMPLE_ACCOUNT_BALANCES} from "../Mocks";
+import {SAMPLE_ACCOUNT, SAMPLE_ACCOUNT_BALANCES} from "../../Mocks";
 
 describe("BalanceAnalyzer.spec.ts", () => {
